chore(vue): Move generate-component to Gulp and add note in generic components.
