+++ bindings/vue/src/components/VOnsActionSheetButton.vue
@@ -5,6 +5,7 @@
 </template>
 
 <script>
+  /* This file is generated automatically */
   import 'onsenui/esm/elements/ons-action-sheet-button';
   import { deriveEvents } from '../mixins';
