+++ bindings/vue/src/components/VOnsAlertDialogButton.vue
@@ -5,6 +5,7 @@
 </template>
 
 <script>
+  /* This file is generated automatically */
   import 'onsenui/esm/elements/ons-alert-dialog-button';
   import { deriveEvents } from '../mixins';
