+++ bindings/vue/src/components/VOnsButton.vue
@@ -5,6 +5,7 @@
 </template>
 
 <script>
+  /* This file is generated automatically */
   import 'onsenui/esm/elements/ons-button';
   import { deriveEvents } from '../mixins';
