+++ bindings/vue/src/components/VOnsCarouselItem.vue
@@ -5,6 +5,7 @@
 </template>
 
 <script>
+  /* This file is generated automatically */
   import 'onsenui/esm/elements/ons-carousel-item';
   import { deriveEvents } from '../mixins';
