+++ bindings/vue/src/components/VOnsCheckbox.vue
@@ -5,6 +5,7 @@
 </template>
 
 <script>
+  /* This file is generated automatically */
   import 'onsenui/esm/elements/ons-checkbox';
   import { deriveEvents, modelCheckbox } from '../mixins';
