+++ bindings/vue/src/components/VOnsCol.vue
@@ -5,6 +5,7 @@
 </template>
 
 <script>
+  /* This file is generated automatically */
   import 'onsenui/esm/elements/ons-col';
   import { deriveEvents } from '../mixins';
