+++ bindings/vue/src/components/VOnsFab.vue
@@ -5,6 +5,7 @@
 </template>
 
 <script>
+  /* This file is generated automatically */
   import 'onsenui/esm/elements/ons-fab';
   import { deriveEvents, hidable } from '../mixins';
