+++ bindings/vue/src/components/VOnsIcon.vue
@@ -5,6 +5,7 @@
 </template>
 
 <script>
+  /* This file is generated automatically */
   import 'onsenui/esm/elements/ons-icon';
   import { deriveEvents } from '../mixins';
