+++ bindings/vue/src/components/VOnsList.vue
@@ -5,6 +5,7 @@
 </template>
 
 <script>
+  /* This file is generated automatically */
   import 'onsenui/esm/elements/ons-list';
   import { deriveEvents } from '../mixins';
