+++ bindings/vue/src/components/VOnsListHeader.vue
@@ -5,6 +5,7 @@
 </template>
 
 <script>
+  /* This file is generated automatically */
   import 'onsenui/esm/elements/ons-list-header';
   import { deriveEvents } from '../mixins';
