+++ bindings/vue/src/components/VOnsListTitle.vue
@@ -5,6 +5,7 @@
 </template>
 
 <script>
+  /* This file is generated automatically */
   import 'onsenui/esm/elements/ons-list-title';
   import { deriveEvents } from '../mixins';
