+++ bindings/vue/src/components/VOnsProgressBar.vue
@@ -5,6 +5,7 @@
 </template>
 
 <script>
+  /* This file is generated automatically */
   import 'onsenui/esm/elements/ons-progress-bar';
   import { deriveEvents } from '../mixins';
