+++ bindings/vue/src/components/VOnsRange.vue
@@ -5,6 +5,7 @@
 </template>
 
 <script>
+  /* This file is generated automatically */
   import 'onsenui/esm/elements/ons-range';
   import { deriveEvents, modelInput } from '../mixins';
