+++ bindings/vue/src/components/VOnsRipple.vue
@@ -5,6 +5,7 @@
 </template>
 
 <script>
+  /* This file is generated automatically */
   import 'onsenui/esm/elements/ons-ripple';
   import { deriveEvents } from '../mixins';
