+++ bindings/vue/src/components/VOnsRow.vue
@@ -5,6 +5,7 @@
 </template>
 
 <script>
+  /* This file is generated automatically */
   import 'onsenui/esm/elements/ons-row';
   import { deriveEvents } from '../mixins';
