+++ bindings/vue/src/components/VOnsSpeedDialItem.vue
@@ -5,6 +5,7 @@
 </template>
 
 <script>
+  /* This file is generated automatically */
   import 'onsenui/esm/elements/ons-speed-dial-item';
   import { deriveEvents } from '../mixins';
