+++ bindings/vue/src/components/VOnsSplitterContent.vue
@@ -5,6 +5,7 @@
 </template>
 
 <script>
+  /* This file is generated automatically */
   import 'onsenui/esm/elements/ons-splitter-content';
   import { deriveEvents } from '../mixins';
