+++ bindings/vue/src/components/VOnsSplitterMask.vue
@@ -5,6 +5,7 @@
 </template>
 
 <script>
+  /* This file is generated automatically */
   import 'onsenui/esm/elements/ons-splitter-mask';
   import { deriveEvents } from '../mixins';
