+++ bindings/vue/src/components/VOnsSwitch.vue
@@ -5,6 +5,7 @@
 </template>
 
 <script>
+  /* This file is generated automatically */
   import 'onsenui/esm/elements/ons-switch';
   import { deriveEvents, modelCheckbox } from '../mixins';
