+++ bindings/vue/src/components/VOnsToast.vue
@@ -5,6 +5,7 @@
 </template>
 
 <script>
+  /* This file is generated automatically */
   import 'onsenui/esm/elements/ons-toast';
   import { deriveEvents, hidable, hasOptions, deriveDBB, portal } from '../mixins';
