+++ bindings/vue/src/components/VOnsToolbar.vue
@@ -5,6 +5,7 @@
 </template>
 
 <script>
+  /* This file is generated automatically */
   import 'onsenui/esm/elements/ons-toolbar';
   import { deriveEvents } from '../mixins';
