+++ bindings/vue/src/components/VOnsToolbarButton.vue
@@ -5,6 +5,7 @@
 </template>
 
 <script>
+  /* This file is generated automatically */
   import 'onsenui/esm/elements/ons-toolbar-button';
   import { deriveEvents } from '../mixins';
